Make validation of input data configurable
By changing the `validators` argument of `lttb.downsample()` from the
default list of validation functions, the data can be checked against
stricter or looser requirements.
For example, if the data is known to always satisfy the requirement for
the x-values to be strictly monotonic, some CPU cycles can be saved by
removing `x_is_strictly_increasing()` from the list.

This change is backwards-compatible.

It would be better to make the `validators` argument keyword-only, but
that would prevent this library from working on Python 2.7.