Add check for NaN values in input data
This check has also been added to the list of default validators.
This change is not backwards-compatible,
but the previous behaviour was that `downsample()` would preferentially
select the NaN values for inclusion in the output
(presumably because it considers these points to be maximally distant
from the anchors in the neighbouring bins),
so users with missing data would either have gotten suboptimal results
(a higher proportion of missing data in the output array)
or they might have implemented their own up-front checks.
So in the unlikely case where this breaks someone's workflow,
this now requires them to make an explicit decision whether to
throw an error, exclude missing values up-front,
or accept output in which nulls are overrepresented.